Use populate_user signal form django_auth_ldap
This saves us a few roundtrips in the ORM, and also makes nasty hacks
juggling nested save() calling unnecessary.

Side effect: We have no way of finding out whether this is the first
login or not this way — I decided that allowing user data to get out
of sync with LDAP is an anti-pattern and removed the feature.